# fast-api-darkside/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/clients":{"get":{"summary":"Get Clients","operationId":"get_clients_clients_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/aaaa"},"type":"array","title":"Response Get Clients Clients Get"}}}}}},"post":{"summary":"Create Client","operationId":"create_client_clients_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/aaaa"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/aaaa"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clients/{client_id}":{"put":{"summary":"Update Client","operationId":"update_client_clients__client_id__put","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"integer","title":"Client Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/leaf"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/leaf"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Client","operationId":"delete_client_clients__client_id__delete","parameters":[{"name":"client_id","in":"path","required":true,"schema":{"type":"integer","title":"Client Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects":{"get":{"summary":"Get Projects","operationId":"get_projects_projects_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Project"},"type":"array","title":"Response Get Projects Projects Get"}}}}}},"post":{"summary":"Create Project","operationId":"create_project_projects_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/{projects_id}":{"put":{"summary":"Update Projects","operationId":"update_projects_projects__projects_id__put","parameters":[{"name":"projects_id","in":"path","required":true,"schema":{"type":"integer","title":"Projects Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Project"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Projects","operationId":"delete_projects_projects__projects_id__delete","parameters":[{"name":"projects_id","in":"path","required":true,"schema":{"type":"integer","title":"Projects Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/members":{"get":{"summary":"Get Members","operationId":"get_members_members_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Member"},"type":"array","title":"Response Get Members Members Get"}}}}}},"post":{"summary":"Create Member","operationId":"create_member_members_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/members/{members_id}":{"put":{"summary":"Update Members","operationId":"update_members_members__members_id__put","parameters":[{"name":"members_id","in":"path","required":true,"schema":{"type":"integer","title":"Members Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Member"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Members","operationId":"delete_members_members__members_id__delete","parameters":[{"name":"members_id","in":"path","required":true,"schema":{"type":"integer","title":"Members Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_slots":{"get":{"summary":"Get Project Slots","operationId":"get_project_slots_project_slots_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectSlot"},"type":"array","title":"Response Get Project Slots Project Slots Get"}}}}}},"post":{"summary":"Create Project Slot","operationId":"create_project_slot_project_slots_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectSlotCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectSlotCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_slots/{project_slot_id}":{"put":{"summary":"Update Project Slot","operationId":"update_project_slot_project_slots__project_slot_id__put","parameters":[{"name":"project_slot_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Slot Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectSlotCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectSlot"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Project Slot","operationId":"delete_project_slot_project_slots__project_slot_id__delete","parameters":[{"name":"project_slot_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Slot Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_budgets":{"get":{"summary":"Get Project Budgets","operationId":"get_project_budgets_project_budgets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectBudget"},"type":"array","title":"Response Get Project Budgets Project Budgets Get"}}}}}},"post":{"summary":"Create Project Budget","operationId":"create_project_budget_project_budgets_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectBudgetCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectBudgetCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_budgets/{project_budget_id}":{"put":{"summary":"Update Project Budget","operationId":"update_project_budget_project_budgets__project_budget_id__put","parameters":[{"name":"project_budget_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Budget Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectBudgetCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectBudget"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Project Budget","operationId":"delete_project_budget_project_budgets__project_budget_id__delete","parameters":[{"name":"project_budget_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Budget Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_member_assign":{"get":{"summary":"Get Project Member Assign","operationId":"get_project_member_assign_project_member_assign_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectMemberAssign"},"type":"array","title":"Response Get Project Member Assign Project Member Assign Get"}}}}}},"post":{"summary":"Create Project Member Assign","operationId":"create_project_member_assign_project_member_assign_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMemberAssignCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMemberAssignCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/project_member_assign/{project_member_assign_id}":{"put":{"summary":"Update Project Member Assign","operationId":"update_project_member_assign_project_member_assign__project_member_assign_id__put","parameters":[{"name":"project_member_assign_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Member Assign Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMemberAssignCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMemberAssign"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Project Member Assign","operationId":"delete_project_member_assign_project_member_assign__project_member_assign_id__delete","parameters":[{"name":"project_member_assign_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Member Assign Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/member_cost":{"get":{"summary":"Get Member Cost","operationId":"get_member_cost_member_cost_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MemberCost"},"type":"array","title":"Response Get Member Cost Member Cost Get"}}}}}},"post":{"summary":"Create Member Cost","operationId":"create_member_cost_member_cost_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCostCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCostCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/member_cost/{member_cost_id}":{"put":{"summary":"Update Member Cost","operationId":"update_member_cost_member_cost__member_cost_id__put","parameters":[{"name":"member_cost_id","in":"path","required":true,"schema":{"type":"integer","title":"Member Cost Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCostCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemberCost"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Member Cost","operationId":"delete_member_cost_member_cost__member_cost_id__delete","parameters":[{"name":"member_cost_id","in":"path","required":true,"schema":{"type":"integer","title":"Member Cost Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/graphql":{"get":{"summary":"Handle Http Get","operationId":"handle_http_get_graphql_get","responses":{"200":{"description":"The GraphiQL integrated development environment.","content":{"application/json":{"schema":{}}}},"404":{"description":"Not found if GraphiQL or query via GET are not enabled."}}},"post":{"summary":"Handle Http Post","operationId":"handle_http_post_graphql_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Member":{"properties":{"name":{"type":"string","title":"Name","example":"山田太郎"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"email":{"type":"string","format":"email","title":"Email","example":"yamada@example.com"},"phone":{"type":"string","title":"Phone","example":"090-1234-5678"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["name","client_id","email","phone","id","created_at","deleted_at"],"title":"Member"},"MemberCost":{"properties":{"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"cost":{"type":"integer","title":"Cost","description":"コスト"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["member_id","start_date","cost","id","created_at","deleted_at"],"title":"MemberCost"},"MemberCostCreate":{"properties":{"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"cost":{"type":"integer","title":"Cost","description":"コスト"}},"type":"object","required":["member_id","start_date","cost"],"title":"MemberCostCreate"},"MemberCostCreateResponse":{"properties":{"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"cost":{"type":"integer","title":"Cost","description":"コスト"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["member_id","start_date","cost","id"],"title":"MemberCostCreateResponse"},"MemberCreate":{"properties":{"name":{"type":"string","title":"Name","example":"山田太郎"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"email":{"type":"string","format":"email","title":"Email","example":"yamada@example.com"},"phone":{"type":"string","title":"Phone","example":"090-1234-5678"}},"type":"object","required":["name","client_id","email","phone"],"title":"MemberCreate"},"MemberCreateResponse":{"properties":{"name":{"type":"string","title":"Name","example":"山田太郎"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"email":{"type":"string","format":"email","title":"Email","example":"yamada@example.com"},"phone":{"type":"string","title":"Phone","example":"090-1234-5678"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","client_id","email","phone","id"],"title":"MemberCreateResponse"},"Project":{"properties":{"name":{"type":"string","title":"Name","example":"プロジェクトA"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"プロジェクト開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"プロジェクト終了日"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["name","client_id","start_date","id","created_at","deleted_at"],"title":"Project"},"ProjectBudget":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["project_id","start_date","budget","id","created_at","deleted_at"],"title":"ProjectBudget"},"ProjectBudgetCreate":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"}},"type":"object","required":["project_id","start_date","budget"],"title":"ProjectBudgetCreate"},"ProjectBudgetCreateResponse":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["project_id","start_date","budget","id"],"title":"ProjectBudgetCreateResponse"},"ProjectCreate":{"properties":{"name":{"type":"string","title":"Name","example":"プロジェクトA"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"プロジェクト開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"プロジェクト終了日"}},"type":"object","required":["name","client_id","start_date"],"title":"ProjectCreate"},"ProjectCreateResponse":{"properties":{"name":{"type":"string","title":"Name","example":"プロジェクトA"},"client_id":{"type":"integer","title":"Client Id","description":"クライアントID"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"プロジェクト開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"プロジェクト終了日"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","client_id","start_date","id"],"title":"ProjectCreateResponse"},"ProjectMemberAssign":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"cost":{"type":"integer","title":"Cost","description":"コスト"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["project_id","member_id","cost","start_date","id","created_at","deleted_at"],"title":"ProjectMemberAssign"},"ProjectMemberAssignCreate":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"cost":{"type":"integer","title":"Cost","description":"コスト"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"}},"type":"object","required":["project_id","member_id","cost","start_date"],"title":"ProjectMemberAssignCreate"},"ProjectMemberAssignCreateResponse":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"member_id":{"type":"integer","title":"Member Id","description":"メンバーID"},"cost":{"type":"integer","title":"Cost","description":"コスト"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"終了日"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["project_id","member_id","cost","start_date","id"],"title":"ProjectMemberAssignCreateResponse"},"ProjectSlot":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"name":{"type":"string","title":"Name","example":"エンジニア"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"募集開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"募集終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At"}},"type":"object","required":["project_id","name","start_date","budget","id","created_at","deleted_at"],"title":"ProjectSlot"},"ProjectSlotCreate":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"name":{"type":"string","title":"Name","example":"エンジニア"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"募集開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"募集終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"}},"type":"object","required":["project_id","name","start_date","budget"],"title":"ProjectSlotCreate"},"ProjectSlotCreateResponse":{"properties":{"project_id":{"type":"integer","title":"Project Id","description":"プロジェクトID"},"name":{"type":"string","title":"Name","example":"エンジニア"},"start_date":{"type":"string","format":"date","title":"Start Date","description":"募集開始日"},"end_date":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"End Date","description":"募集終了日"},"budget":{"type":"integer","title":"Budget","description":"予算"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["project_id","name","start_date","budget","id"],"title":"ProjectSlotCreateResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"aaaa":{"properties":{"name":{"type":"string","maxLength":255,"title":"Name"}},"additionalProperties":false,"type":"object","required":["name"],"title":"aaaa"},"leaf":{"properties":{"id":{"type":"integer","maximum":9.223372036854776e+18,"minimum":1.0,"title":"Id"},"name":{"type":"string","maxLength":255,"title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At","readOnly":true},"deleted_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Deleted At","nullable":true}},"additionalProperties":false,"type":"object","required":["id","name","created_at","deleted_at"],"title":"Client"}}}}